requirements.txt
torch==1.9.0
torchvision==0.10.0
numpy==1.21.2
tqdm==4.62.3
Pillow==8.3.2
